safari-600.7-branch/LayoutTests/fast/dom/node-legacy-event-listener-expected.txt

Test addEventListener() and removeEventListener() fail silently if arguments are missing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.addEventListener("foo") is undefined
PASS document.removeEventListener("bar") is undefined
PASS document.addEventListener() is undefined
PASS document.removeEventListener() is undefined
PASS successfullyParsed is true

TEST COMPLETE
